Allows submission on enter
In order to get enter to work I had to change the element to a form and
give that form an onSubmit event listener.  Then the function it
triggers is an anonymous function that has access to the event.  In that
anon function I call my handleSubmit with the event in order to use
preventDefault() from automatically refreshing on submit.